refactor: problem -> article 패키지 이동
